Split of service.go into msg/service.go
Create a new package called 'msg', just like for SkyDNS1. This
is need because some transformation function will be used
outside SkyDNS.

The methods names in there need to be cleaned up slightly.
